requirements.txt
transformers>=4.44
accelerate
torch
numpy
pandas
pyyaml
tqdm
